This adds a "scheduler" output to the reporting.
It's a bit more involved since we have to keep the last
scheduler_wall_time result around to diff it against the current one,
as per recon's scheduler_usage function. The natural place to keep
this state seems to be the ehmon_report_srv gen_server.

Upon every triggering of the timeout, we read scheduler wall clock
stats and compare the values with those from the last run. The value
for each core gets diffed with the previous one, and we divide the
active time by total time to get a rough percentage of scheduled
time. This is summed up and emitted in the report line.