Expose raw proplist to report_mf callback.
If the report_mf app config value is set to a tuple of {M,F,raw}, then
it will be called with the raw proplist of values rather than the
stringified version. This will allow consumers to do things like emit
lines in l2met-friendly formats.

I'm not totally sure about this method of indicating you want access to
the raw values; maybe a separate config should be used?