implement submitting reports to shh
Using either `ehmon_shh_socket` or the combined configuration options of
`ehmon_shh_host` and `ehmon_shh_port` it is now possible for ehmon to
report erlang vm metrics to a listening shh process.

Connections to shh will automatically be re-established if lost. Any
received reports while a connection is down will be queued up and sent
when the connection is reestablished.