make gen_unix a soft dependency
The gen_unix library is not used by the default configuration. It is
only required when a user wants to connect to shh via it's listening
unix socket.

A note has been added to the README.rst file noting this soft
requirement.

When a user attempts to configure ehmon to communicate over unix socket
and the gen_unix app is not available the following standard error will
be displayed.

```erlang
1> {"init terminating in do_boot",{{app_start_failed,gen_unix,{"no such file or directory","gen_unix.app"}},[{ehmon_app,start_ok,3,[{file,"src/ehmon_app.erl"},{line,67}]},{ehmon_app,start_ok,3,[{file,"src/ehmon_app.erl"},{line,64}]},{init,start_it,1,[]},{init,start_em,1,[]}]}}

    Crash dump was written to: erl_crash.dump
    init terminating in do_boot ()
  ```